+++ smallworld/emulators/unicorn.py
@@ -113,8 +113,7 @@ class UnicornEmulator(emulator.Emulator):
         if self.mode not in self.REGISTERS[self.arch]:
             raise ValueError("unsupported mode for current architecture")
 
-        self.entrypoint: typing.Optional[int] = None
-        self.exitpoint: typing.Optional[int] = None
+        self.entry: typing.Optional[int] = None
 
         self.engine = unicorn.Uc(self.arch, self.mode)
 
@@ -122,7 +121,7 @@ class UnicornEmulator(emulator.Emulator):
             self.CAPSTONE_ARCH_MAP[self.arch], self.CAPSTONE_MODE_MAP[self.mode]
         )
         self.disassembler.detail = True
-        self.pc_ranges: typing.List[range] = []
+        self.bounds: typing.Iterable[range] = []
 
         self.hooks: typing.Dict[int, typing.Callable[[emulator.Emulator], None]] = {}
         self.hook_return = None
@@ -254,29 +253,17 @@ class UnicornEmulator(emulator.Emulator):
         self.write_memory(code.base, code.image)
 
         if code.entry is not None:
-            self.entrypoint = code.entry
-            if self.entrypoint < code.base or self.entrypoint > code.base + len(
-                code.image
-            ):
+            self.entry = code.entry
+            if self.entry < code.base or self.entry > code.base + len(code.image):
                 raise ValueError(
-                    "entrypoint is not in code: 0x{self.entrypoint:x} vs (0x{code.base:x}, 0x{code.base + len(code.image):x})"
+                    "entry is not in code: 0x{self.entry:x} vs (0x{code.base:x}, 0x{code.base + len(code.image):x})"
                 )
         else:
-            self.entrypoint = code.base
+            self.entry = code.base
 
-        if code.exits:
-            self.exitpoint = list(code.exits)[0]
-        else:
-            self.exitpoint = code.base + len(code.image)
-
-        if len(self.pc_ranges) == 0:
-            pcr = range(self.entrypoint, self.exitpoint)
-            self.pc_ranges.append(pcr)
+        self.bounds = code.bounds
 
         logger.info(f"loaded code (size: {len(code.image)} B) at 0x{code.base:x}")
-
-    def add_pc_range(self, pc_range: range) -> None:
-        self.pc_ranges.append(pc_range)
 
     def hook(
         self, address: int, function: typing.Callable[[emulator.Emulator], None]
@@ -320,24 +307,36 @@ class UnicornEmulator(emulator.Emulator):
         for i in self.disassembler.disasm(code, pc):
             return i
 
+    @property
+    def exit(self):
+        for bound in self.bounds:
+            if self.entry in bound:
+                return bound.stop
+
+        return None
+
     def check(self) -> None:
-        if self.entrypoint is None:
+        if self.entry is None:
             raise exceptions.ConfigurationError(
-                "no entrypoint provided, emulation cannot start"
+                "no entry provided, emulation cannot start"
             )
-        if self.exitpoint is None:
+
+        if not self.bounds:
             raise exceptions.ConfigurationError(
-                "no exitpoint provided, emulation cannot start"
+                "no bounds provided, emulation cannot start"
+            )
+
+        if self.exit is None:
+            raise exceptions.ConfigurationError(
+                "entry is not in valid execution bounds"
             )
 
     def run(self) -> None:
         self.check()
 
-        logger.info(
-            f"starting emulation at 0x{self.entrypoint:x} until 0x{self.exitpoint:x}"
-        )
+        logger.info(f"starting emulation at 0x{self.entry:x} until 0x{self.exit:x}")
         try:
-            self.engine.emu_start(self.entrypoint, self.exitpoint)
+            self.engine.emu_start(self.entry, self.exit)
         except unicorn.UcError as e:
             logger.warn(f"emulation stopped - reason: {e}")
             logger.warn("for more details, run emulation in single step mode")
@@ -358,22 +357,18 @@ class UnicornEmulator(emulator.Emulator):
         logger.info(f"single step at 0x{pc:x}: {disas}")
 
         try:
-            self.engine.emu_start(pc, self.exitpoint, count=1)
+            self.engine.emu_start(pc, self.exit, count=1)
         except unicorn.UcError as e:
             logger.warn(f"emulation stopped - reason: {e}")
             self._error(e)
 
         pc = self.read_register("pc")
 
-        # if we have provided any pc ranges, then we are using
-        # them and nothing else to determine when we are done
-        # ** if pc is *in* any of the pc ranges in our list then we are not done.
-        done = True
-        for pc_range in self.pc_ranges:
-            if pc in pc_range:
-                done = False
-                break
-        return done
+        for bound in self.bounds:
+            if pc in bound:
+                return False
+
+        return True
 
     def _error(
         self, error: unicorn.UcError